Features #5321: Added robustness features to Monitoring. Also renamed the packet loss monitoring point in the ingest pipeline.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@5440 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/CP/ingest/current/ingestpipeline/IngestPipeline.cc
@@ -83,7 +83,9 @@ void IngestPipeline::ingest(void)
     const std::vector<TaskDesc>& tasks = itsConfig.tasks();
 
     // 2) Configure the Monitoring Singleton
-    MonitoringSingleton::init(itsConfig);
+    if (!itsConfig.monitoringArchiverService().registryHost().empty()) {
+        MonitoringSingleton::init(itsConfig);
+    }
 
     // 3) Create a Task Factory
     TaskFactory factory(itsConfig);
@@ -119,7 +121,9 @@ void IngestPipeline::ingest(void)
 
     // 7) Clean up
     itsSource.reset();
-    MonitoringSingleton::destroy();
+    if (MonitoringSingleton::instance()) {
+        MonitoringSingleton::destroy();
+    }
 }
 
 bool IngestPipeline::ingestOne(void)
